Do not make tracker-indexer commit suicide if tracker-extract doesn't get along.
If tracker-extract fails to return the PID. extractor_pid is 0, so if
tracker-indexer tries to kill it afterwards, it ends up killing itself.

+++ src/tracker-indexer/tracker-module-metadata-utils.c
@@ -288,7 +288,12 @@ extractor_context_destroy (ExtractorContext *context)
 static void
 extractor_context_kill (ExtractorContext *context)
 {
-        g_message ("Attempting to kill tracker-extract with SIGKILL");
+	g_message ("Attempting to kill tracker-extract with SIGKILL");
+
+	if (context->pid == 0) {
+		g_warning ("  No PID for tracker-extract");
+		return;
+	}
 
         if (kill (context->pid, SIGKILL) == -1) {
                 const gchar *str = g_strerror (errno);
